// frontend-v2/src/app/core/api/models/adicional-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AdicionalItemOutput } from '../models/adicional-item-output';

/**
 * Dados de saida do grupo de adicionais
 */
export interface AdicionalOutput {

  /**
   * ID do adicional
   */
  id_adicional?: number;

  /**
   * ID da categoria
   */
  id_categoria?: number;

  /**
   * Limite (quando aplicavel)
   */
  limite?: number;

  /**
   * Minimo (quando aplicavel)
   */
  minimo?: number;

  /**
   * Nome do grupo
   */
  nome?: string;

  /**
   * Itens de adicional
   */
  opcoes?: Array<AdicionalItemOutput>;

  /**
   * Tipo de selecao (U, M, Q)
   */
  selecao?: string;

  /**
   * Status (ativo/inativo)
   */
  status?: boolean;
}